CLI reintegration now works for illumina
